Update Foundation projects; fix issue with Wikipedia namespace.
Please, update the current Foundation projects.
There were also 2 typos: "speciesi" and "meta-wiki".
We probably don't need these:

tools   bugzilla   mediazilla   quality   strategy   usability   
betawikiversity   oldwikisource   nost   nostalgia   tenwiki   donate

If possible, do not add "w" and "wikipedia" until errors #68 and #82 are rewritten to take care of them, as explained here:

https://en.wikipedia.org/wiki/Wikipedia_talk:WikiProject_Check_Wikipedia#Error_82_confusion_and_new_error_104

For the same reason, "wikt" and "wiktionary" are special cases when the script is running for sv.wiktionary; and so on with other exceptions when running for commons, etc. At the moment, errors  #68 and #82 are not active in those projects, so my fix doesn't have effects on them.

As explained in the link above, "w" and "wikipedia" can also combine in multiple ways with language codes and with "m", "mw" and the like, with the script unable to sort out all the cases at the moment; see how the script works in es.wiki, where errors #68 and #82 have the majority of the problems (my fix doesn't solve all of them; for example it can't stop [[:de:wikt:Kreis|kreis]] showing up in error #68 instead of #82 ).